campus_connect/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
gunicorn==21.2.0
inflection==0.5.1
packaging==23.2
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
